Salon FBX + Textures
Here is the updated salon fbx and textures without lights. Its split up into 3 files but hopefully it works out. Everything should work fine in unity i tried it out on a personal file. Just make sure to plug in the normal or metallic maps for the objects that have them. If you don't know how to do that i can show you.